Move ModifierKeys into own file modifier_keys.rs
